requirements.txt
blinker==1.8.2
cffi==1.16.0
click==8.1.7
colorama==0.4.6
cryptography==42.0.8
distlib==0.3.8
filelock==3.15.4
Flask==3.0.3
Flask-Cors==4.0.1
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
platformdirs==4.2.2
psycopg2-binary==2.9.9
pycparser==2.22
SQLAlchemy==2.0.31
typing_extensions==4.12.2
virtualenv==20.26.3
Werkzeug==3.0.3
